route_planner/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
flask>=2.0.0
googlemaps>=4.10.0
folium>=0.14.0
geopy>=2.4.0 